Move out tunes and instruments to separate files
